refactor: single source of truth for skills quiz course cards
